+++ .github/workflows/ci_gradle.yml
@@ -25,7 +25,7 @@ jobs:
           echo "  profiles:" >> application.yml
           echo "    group:" >> application.yml
           echo "      "local": "db, s3, secret"" >> application.yml
-          echo "      "prod": "db, s3"" >> application.yml
+          echo "      "prod": "db, s3, redis"" >> application.yml
           echo "    default: prod" >> application.yml
           
           echo "  servlet:" >> application.yml
@@ -49,8 +49,6 @@ jobs:
           spring.jwt.secret: ${{ secrets.JWT_SECRET }}
           spring.jwt.token.access-expiration-time: ${{ secrets.JWT_ACCESS_EXPIRATION_TIME }}
           spring.jwt.token.refresh-expiration-time: ${{ secrets.JWT_REFRESH_EXPIRATION_TIME }}
-          spring.data.redis.host: ${{ secrets.REDIS_HOST }}
-          spring.data.redis.port: ${{ secrets.REDIS_PORT }}
           mail.smtp.port: ${{ secrets.EMAIL_PORT }}
           AdminMail.id: ${{ secrets.EMAIL_ID }}
           AdminMail.password: ${{ secrets.EMAIL_PASSWORD }}
@@ -77,6 +75,14 @@ jobs:
           cloud.aws.region.static: ${{ secrets.REGION }}
           cloud.aws.credentials.accessKey: ${{ secrets.S3_ACCESS_KEY }}
           cloud.aws.credentials.secretKey: ${{ secrets.S3_SECRET_KEY }}
+
+      - name: 6) Set application-redis.yml
+        uses: microsoft/variable-substitution@v1
+        with:
+          files: ./core/core-infra-redis/src/main/resources/application-redis.yml
+        env:
+          spring.data.redis.host: ${{ secrets.REDIS_HOST }}
+          spring.data.redis.port: ${{ secrets.REDIS_PORT }}
 
       - name: Grant execute permission for gradlew and build
         run: |
