+++ core/core-domain/src/main/java/com/sponus/CoreDomainConfig.java
@@ -4,14 +4,12 @@ import org.springframework.boot.autoconfigure.domain.EntityScan;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
 import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
-import org.springframework.data.redis.repository.configuration.EnableRedisRepositories;
 import org.springframework.transaction.annotation.EnableTransactionManagement;
 
 @Configuration
 @EnableTransactionManagement
 @EntityScan("com.sponus.coredomain")
 @EnableJpaRepositories("com.sponus.coredomain")
-@EnableRedisRepositories("com.sponus.coredomain")
 @EnableJpaAuditing
 public class CoreDomainConfig {
 }
