+++ core/core-infra-redis/build.gradle
@@ -1,0 +1,8 @@
+dependencies {
+    // Redis
+    implementation group: 'org.springframework.boot', name: 'spring-boot-starter-data-redis', version: '3.2.0'
+    implementation 'org.springframework.session:spring-session-data-redis:3.1.1'
+}
+
+bootJar { enabled = false }
+jar { enabled = true }
