+++ core/core-infra-redis/src/main/java/com/sponus/coreinfraredis/CoreRedisConfig.java
@@ -1,0 +1,9 @@
+package com.sponus.coreinfraredis;
+
+import org.springframework.context.annotation.Configuration;
+import org.springframework.data.redis.repository.configuration.EnableRedisRepositories;
+
+@Configuration
+@EnableRedisRepositories("com.sponus.coreinfraredis")
+public class CoreRedisConfig {
+}
