+++ core/core-infra-redis/src/main/java/com/sponus/coreinfraredis/entity/AnnouncementView.java
@@ -1,4 +1,4 @@
-package com.sponus.coredomain.domain.announcement;
+package com.sponus.coreinfraredis.entity;
 
 import java.time.Duration;
 import java.time.LocalDate;
