+++ core/core-infra-redis/src/main/resources/application-redis.yml
@@ -1,0 +1,5 @@
+spring:
+  data:
+    redis:
+      host: ${REDIS_HOST}
+      port: ${REDIS_PORT}
