+++ settings.gradle
@@ -3,3 +3,4 @@ include 'api'
 include 'core:core-domain'
 include 'core:core-infra-db'
 include 'core:core-infra-s3'
+include 'core:core-infra-redis'
